stt/requirements.txt
Flask==2.3.2
google-cloud-speech==2.21.1
requests
mysql-connector-python
pandas